requirements.txt
appdirs==1.4.2
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pyparsing==2.1.10
pyTelegramBotAPI==2.3.0
requests==2.13.0
six==1.10.0
Werkzeug==0.11.15
psycopg2==2.7
Flask-SQLAlchemy==2.2
